# swagger.yaml
openapi: 3.0.0
info:
  title: 'PDB API'
  description: 'API documentation.'
  version: '1.0'
servers:
  -
    url: 'https://pdb.local'
    description: Dev
  -
    url: 'https://localhost'
    description: Local
paths:
...
  /command/material:
    post:
      tags:
        - command
        - material
      description: 'Create a new material.'
      operationId: createMaterial
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreateRequest'
      responses:
        '201':
          $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  '/command/material/{id}':
    put:
      tags:
        - command
        - material
      description: 'Update a material.'
      operationId: updateMaterial
      parameters:
        -
          name: id
          in: path
          description: 'Material ID'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialUpdateRequest'
      responses:
        '200':
          $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - command
        - material
      description: 'Delete a material.'
      operationId: deleteMaterial
      parameters:
        -
          name: id
          in: path
          description: 'Material ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/schemas/ErrorResponse'
        '412':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MaterialCreateRequest:
      properties:
        name:
          type: string
        size:
          type: integer
          minimum: 1
        wildcard:
          description: 'Used in processes where concrete materials do not matter'
          type: boolean
      type: object
...
    MaterialUpdateRequest:
      properties:
        name:
          type: string
          nullable: true
        size:
          type: integer
          nullable: true
        wildcard:
          type: boolean
          nullable: true
      type: object
...
    ErrorResponse:
      properties:
        title:
          type: string
          example: 'Error title'
        type:
          $ref: '#/components/schemas/ErrorType'
        detail:
          type: string
          example: 'Error description'
          nullable: true
        exception:
          properties:
            message:
              type: string
              example: 'Some exception description'
            code:
              type: integer
              example: 123
            trace:
              type: array
              items:
                type: object
              example:
                -
                  file: index.php
                  line: 1
                  function: abc
                  args: ['Argument value']
          type: object
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
    SuccessResponse:
      properties:
        message:
          type: string
          example: Message
        detail:
          type: string
          example: Description
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
    ErrorType:
      type: string
      enum:
        - validation_error
        - database_error
        - internal_error
        - resource_not_found_error
        - resource_access_error
...